requirements.txt
numpy
matplotlib
pytorch==1.6.0
pytorch-lightning==1.0.2
tqdm
